Fix some stuff, introduce DataPreparer module that handles data preparation, limits, condensing, sort etc.
